+++ src/main/java/com/example/ecommercea/service/SellerService.java
@@ -26,6 +26,7 @@ public class SellerService {//판매자
         //판매자 인증 - 소비자가 로그인을 하려고 할 때, 소비자 이름, PW 확인한 후에 로그인
         Seller seller= sellerRepository.findBySellerName
                 (sellerLoginRequest.getSellerName());
+
         if(seller.getSellerPW().equals(sellerLoginRequest.getSellerPW())
                 &&seller.getSellerName().equals(sellerLoginRequest.getSellerName()))
         {return seller;}return null;}
@@ -43,11 +44,16 @@ public class SellerService {//판매자
     public Seller changeSeller(Long sellerID, SellerRegister sellerRegister){
         Seller saved= sellerRepository.findById(sellerID)
                 .orElseThrow(()->new RuntimeException("판매자 없음"));
-        saved.sellerChange(sellerRegister);return saved;}
+
+        if(sellerRepository.existsBySellerName(sellerRegister.getSellerName())){
+            throw new DuplicateNameException("판매자 이름 이미 있음");}
+        saved.sellerChange(sellerRegister);return sellerRepository.save(saved);}//[4][31][32][33][34]
 
     public void deleteSeller(Long sellerID, SellerDeleteRequest sellerDeleteRequest){
         Seller seller= sellerRepository.findById(sellerID)
                 .orElseThrow(()->new RuntimeException("판매자 없음"));
-        if(seller.getSellerPW().equals(sellerDeleteRequest.getSellerPW())){
+
+        if(seller.getSellerPW().equals(sellerDeleteRequest.getSellerPW())&&
+        seller.getSellerName().equals(sellerDeleteRequest.getSellerName())){
             sellerRepository.deleteById(sellerID);}
-        else throw new RuntimeException("판매자 ID 삭제 실패 또는 발견 안 됨");}}
+        else throw new RuntimeException("판매자 ID 삭제 실패 또는 발견 안 됨");}}//[5][36][37][38]
